+++ packages/abis/index.ts
@@ -1,4 +1,4 @@
-export * from './FollowNfts';
+export * from './FollowNft';
 export * from './LensHub';
 export * from './LensPeriphery';
 export * from './UpdateOwnableFeeCollectModule';
